requirements.txt
ansible==8.3.0
ansible-builder==3.0.0
ansible-core==2.15.3
molecule==6.0.2
molecule-plugins==23.5.0
pytest-testinfra==9.0.0
